Trading_bot/_archive/task1/new_listings_parser/server_version/pythonProject_copy/requirements.txt
requests
datetime
gate_api
beautifulsoup4
numpy
pybit
asyncio
urllib3
python-telegram-bot>10.0
jsonref
python-okx